separate ProjectView from ConsoleDisplay
