+++ src/main/kotlin/ui/views/project/ProjectView.kt
@@ -1,0 +1,15 @@
+package org.damascus.ui.views.project
+
+import logic.model.Project
+
+interface ProjectView {
+
+    fun displayAllProjects(
+        projects: List<Project>,
+        label: String,
+        contentSelector: ((Project) -> Map<String, Any?>)? = null
+    )
+
+    fun displayProjectsAsTable(projects: List<Project>)
+
+}
